requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
fastapi-mail==1.4.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.1
alembic==1.12.1
python-dotenv==1.0.0
jinja2==3.1.2
pytest==7.4.3
httpx==0.25.1
asyncpg==0.28.0
SQLAlchemy==2.0.23
pyjwt==2.8.0
bcrypt==4.0.1
aiosmtplib==2.0.2
slowapi==0.1.8